chore: renamed openai store
